pyproject.toml
# MLOps Obesity Classifier Configuration

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mlops-obesity-classifier"
version = "1.0.0"
description = "Production-ready MLOps pipeline for obesity classification"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Alicia Canta", email = "alicia.cantarero@example.com" }]
maintainers = [
    { name = "Alicia Canta", email = "alicia.cantarero@example.com" },
]
keywords = [
    "mlops",
    "machine-learning",
    "obesity-classification",
    "healthcare",
    "mlflow",
    "dvc",
    "production-ml",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "mlflow>=2.5.0",
    "dvc>=3.20.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pyyaml>=6.0",
    "flask>=2.3.0",
    "scipy>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.3",
]
docs = ["sphinx>=7.1.0", "sphinx-rtd-theme>=1.3.0", "myst-parser>=2.0.0"]
monitoring = ["prometheus-client>=0.17.1", "grafana-api>=1.0.3"]

[project.urls]
Homepage = "https://github.com/yourusername/mlops-reproducible"
Documentation = "https://github.com/yourusername/mlops-reproducible/docs"
Repository = "https://github.com/yourusername/mlops-reproducible"
"Bug Tracker" = "https://github.com/yourusername/mlops-reproducible/issues"

[project.scripts]
mlops-train = "src.scripts.train:main"
mlops-predict = "src.scripts.predict:main"
mlops-monitor = "src.scripts.monitor:main"
mlops-serve = "src.scripts.serve_model:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json", "*.md"]

# Black Configuration
[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# isort Configuration  
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100
known_first_party = ["src"]

# MyPy Configuration
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

# Pytest Configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# Coverage Configuration
[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "setup.py", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]
